Update README.md
what i did was i put out codes that will make a sqaure but more sqaure inside,so then i tried differnt code to make it and after like 5 minutes i got it.SO then i copy that over and over and over and made a bigger square. i had to add a little detial which were the eyes so then went back to fix it so i can put in eye, in 2 minutes i did it and got this. Thank you